compress: Update lz4 implementation to new API
As of LZ4 v1.7, the LZ4_compress() function is deprecated and
is to be replaced with LZ4_compress_default().

This new function adds an extra argument which declares the size
of the destination buffer.  In addition, if the result is 0, the
result buffer is not to be trusted as the compression failed.

Signed-off-by: David Sommerseth <davids@openvpn.net>